Linked list, symbolic lightshows
I’ve made big changes to the architecture.  The Master List is gone.
There is now a Symbolic Lightshow list that holds lightshows that
schedule other lightshows.  The Primitive Lightshow list is now a
linked list.  Because of these changes the spreadsheet translator
project is out of date.  The automatically generated code is not
compatible until that project is updated.